+++ jgclark.Dashboard/src/pluginToHTMLBridge.js
@@ -20,7 +20,7 @@ import {
   doCyclePriorityStateUp,
   doDeleteItem,
   doEvaluateString,
-  doDashboardSettingsChanged,
+  doSaveDashboardSettingsFromBridge,
   doShowNoteInEditorFromFilename,
   doShowNoteInEditorFromTitle,
   doShowLineInEditorFromFilename,
@@ -37,7 +37,7 @@ import {
   doRenamePerspective,
   doSavePerspective,
   doSwitchToPerspective,
-  doPerspectiveSettingsChanged,
+  doSavePerspectiveSettingsFromBridge,
 } from './perspectiveClickHandlers'
 import { incrementallyRefreshSomeSections, refreshSectionsBatch, refreshSomeSections } from './refreshClickHandlers'
 import {
@@ -341,11 +341,11 @@ export async function bridgeClickDashboardItem(data: MessageDataObject) {
         break
       }
       case 'dashboardSettingsChanged': {
-        result = await doDashboardSettingsChanged(data, 'dashboardSettings')
+        result = await doSaveDashboardSettingsFromBridge(data, 'dashboardSettings')
         break
       }
       case 'perspectiveSettingsChanged': {
-        result = await doPerspectiveSettingsChanged(data)
+        result = await doSavePerspectiveSettingsFromBridge(data)
         break
       }
       case 'addNewPerspective': {
@@ -785,8 +785,8 @@ async function processActionOnReturn(handlerResultIn: TBridgeClickHandlerResult,
     }
 
     if (actionsOnSuccess.includes('START_DELAYED_REFRESH_TIMER')) {
-      // TEST: turning this off for now
-      logDebug('processActionOnReturn', `START_DELAYED_REFRESH_TIMER: 😳 NOT NOW setting startDelayedRefreshTimer in pluginData`)
+      // TEST: turning this off for now: make it INFO level so I see it
+      logInfo('processActionOnReturn', `START_DELAYED_REFRESH_TIMER: 😳 NOT NOW setting startDelayedRefreshTimer in pluginData`)
       // const reactWindowData = await getGlobalSharedData(WEBVIEW_WINDOW_ID)
       // reactWindowData.pluginData.startDelayedRefreshTimer = true
       // await sendToHTMLWindow(WEBVIEW_WINDOW_ID, 'UPDATE_DATA', reactWindowData, `Setting startDelayedRefreshTimer`)